PHEME dataset models, parser for tweets and structure json files and graph builder
some TODOs in the feature extraction (especially the text analysis and the temporal encoding need to be done.)

Unit tests use pytest and are located in the tests directory.
Integration tests are in the root dir for now, confirmed to parse all tweets in the dataset without errors.  Warnings are generated at the default debug level when we find that the curated "structure" from PHEME does not contain all the full reply structure.  not clear which of the two we want to use for the GAT, so including both.

PHEME dataset should be extracted in the data/ directory, which is includeed in the .gitignore file to minimize noise and repo size.