Adding tests and examples.
Pydantic used for json validation, pytest for unit tests.

Fields which are marked optional are not found in all of the tweet json entries.

test_all_tweets.py can be used to confirm that all tweets in the PHEME dataset (which should be placed in the data/ directory are parsed correctly.